+++ inmemorydb.h
@@ -18,15 +18,15 @@ namespace sus
 
             bool createNewsgroup(const string& name);
 
-            bool deleteNewsgroup(unsigned long newsIdent);
+            bool deleteNewsgroup(unsigned int newsIdent);
 
-            bool existsNewsgroup(unsigned long newsIdent);
+            bool existsNewsgroup(unsigned int newsIdent);
 
-            vector<Article> getArticles(unsigned long newsIdent);
-            bool createArticle(unsigned long newsIdent, const string& title,
+            vector<Article> getArticles(unsigned int newsIdent);
+            bool createArticle(unsigned int newsIdent, const string& title,
                 const string& author, const string& body);
-            bool deleteArticle(unsigned long newsIdent, unsigned long artIdent);
-            Article getArticle(unsigned long newsIdent, unsigned long artIdent);
+            bool deleteArticle(unsigned int newsIdent, unsigned int artIdent);
+            Article getArticle(unsigned int newsIdent, unsigned int artIdent);
         private:
             vector<MemoryNewsgroup> newsgroups;
     };
