+++ susdb.h
@@ -12,13 +12,13 @@ namespace sus
         string title;
         string author;
         string body;
-        unsigned long ident;
+        unsigned int ident;
     };
 
     struct Newsgroup
     {
         string name;
-        unsigned long ident;
+        unsigned int ident;
     };
 
     class DatabaseInterface
@@ -34,24 +34,24 @@ namespace sus
             virtual bool createNewsgroup(const string& name) = 0;
 
             /* Delete newsgroup, return false if not exist */
-            virtual bool deleteNewsgroup(unsigned long newsIdent) = 0;
+            virtual bool deleteNewsgroup(unsigned int newsIdent) = 0;
 
             /* Return true if newsgroup exist */
-            virtual bool existsNewsgroup(unsigned long newsIdent) = 0;
+            virtual bool existsNewsgroup(unsigned int newsIdent) = 0;
 
             /* Return list of articles in newsgroup */
-            virtual vector<Article> getArticles(unsigned long newsIdent) = 0;
+            virtual vector<Article> getArticles(unsigned int newsIdent) = 0;
 
             /* Create article in newsgroup identified by newsIdent
              * Return false of newsgroup does not exist */
-            virtual bool createArticle(unsigned long newsIdent, const string& title,
+            virtual bool createArticle(unsigned int newsIdent, const string& title,
                 const string& author, const string& body) = 0;
 
             /* Delete article in newsident, return false if not possible */
-            virtual bool deleteArticle(unsigned long newsIdent, unsigned long artIdent) = 0;
+            virtual bool deleteArticle(unsigned int newsIdent, unsigned int artIdent) = 0;
 
             /* Return article */
-            virtual Article getArticle(unsigned long newsIdent, unsigned long artIdent) = 0;
+            virtual Article getArticle(unsigned int newsIdent, unsigned int artIdent) = 0;
     };
 }
